+++ NHibernate.JustBlog.Core/IBlogRepository.cs
@@ -1,0 +1,11 @@
+﻿using NHibernate.JustBlog.Core.Models;
+using System.Collections.Generic;
+
+namespace NHibernate.JustBlog.Core
+{
+    public interface IBlogRepository
+    {
+        IList<Post> Posts(int pageNo, int pageSize);
+        int TotalPosts();
+    }
+}